# app/Http/Resources/TabResourceSchema.php
<?php

declare(strict_types=1);

namespace App\Http\Resources;

use OpenApi\Attributes as OA;

#[OA\Schema(
    schema: 'TabResource',
    properties: [
        new OA\Property(property: 'id', type: 'string', format: 'uuid'),
        new OA\Property(property: 'title', type: 'string'),
        new OA\Property(property: 'position', type: 'integer'),
        new OA\Property(property: 'last_note_at', type: 'string', format: 'date-time', nullable: true),
        new OA\Property(property: 'notes_count', type: 'integer'),
        new OA\Property(property: 'created_at', type: 'string', format: 'date-time', nullable: true),
        new OA\Property(property: 'updated_at', type: 'string', format: 'date-time', nullable: true),
    ],
)]
final class TabResourceSchema {}