# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Note API",
        "description": "JSON API для вкладок и заметок",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API"
        }
    ],
    "paths": {
        "/tabs/{tab}/notes": {
            "get": {
                "tags": [
                    "Notes"
                ],
                "summary": "Список заметок вкладки",
                "operationId": "b9cb4f99f8768ad43ec859fe4bda7c11",
                "parameters": [
                    {
                        "name": "tab",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "kind",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "regular",
                                "important",
                                "trash"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/NoteResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Notes"
                ],
                "summary": "Создать заметку",
                "operationId": "f9069461d58078d3abe1bf382ea5b4c3",
                "parameters": [
                    {
                        "name": "tab",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "nullable": true,
                                        "maxLength": 120
                                    },
                                    "body": {
                                        "type": "string",
                                        "nullable": true,
                                        "maxLength": 20000
                                    },
                                    "kind": {
                                        "type": "string",
                                        "enum": [
                                            "regular",
                                            "important",
                                            "trash"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/NoteResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/notes/{note}": {
            "delete": {
                "tags": [
                    "Notes"
                ],
                "summary": "Удалить заметку (в корзину или окончательно из корзины)",
                "operationId": "972df6cf3631054ea9841ffd46a82a03",
                "parameters": [
                    {
                        "name": "note",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Notes"
                ],
                "summary": "Обновить заметку",
                "operationId": "e15d1833d82c5834b4ee56df0ee7323b",
                "parameters": [
                    {
                        "name": "note",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "nullable": true,
                                        "maxLength": 120
                                    },
                                    "body": {
                                        "type": "string",
                                        "nullable": true,
                                        "maxLength": 20000
                                    },
                                    "kind": {
                                        "type": "string",
                                        "enum": [
                                            "regular",
                                            "important",
                                            "trash"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/NoteResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tabs": {
            "get": {
                "tags": [
                    "Tabs"
                ],
                "summary": "Список вкладок",
                "operationId": "affd60c9019e7966bf47ca735ddbcf38",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/TabResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tabs"
                ],
                "summary": "Создать вкладку",
                "operationId": "40881c3e47f0ae30257de3bd89b84670",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "nullable": true,
                                        "maxLength": 120
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/TabResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tabs/{tab}": {
            "delete": {
                "tags": [
                    "Tabs"
                ],
                "summary": "Удалить вкладку",
                "operationId": "8f1cd1a6ec0f2291235bcbbe266abb91",
                "parameters": [
                    {
                        "name": "tab",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Tabs"
                ],
                "summary": "Обновить вкладку",
                "operationId": "095698682d1182f11f9cc5817a464b69",
                "parameters": [
                    {
                        "name": "tab",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "maxLength": 120
                                    },
                                    "position": {
                                        "type": "integer",
                                        "minimum": 0
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/TabResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tabs/reorder": {
            "patch": {
                "tags": [
                    "Tabs"
                ],
                "summary": "Изменить порядок вкладок",
                "operationId": "88fd4d2395b148cea5f952d54d0ecb1c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "ids"
                                ],
                                "properties": {
                                    "ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "uuid"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/TabResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "NoteResource": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "tab_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "title": {
                        "type": "string",
                        "nullable": true
                    },
                    "body": {
                        "type": "string",
                        "nullable": true
                    },
                    "kind": {
                        "type": "string",
                        "enum": [
                            "regular",
                            "important",
                            "trash"
                        ]
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "TabResource": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "title": {
                        "type": "string"
                    },
                    "position": {
                        "type": "integer"
                    },
                    "last_note_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "notes_count": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Notes",
            "description": "Notes"
        },
        {
            "name": "Tabs",
            "description": "Tabs"
        }
    ]
}